Allow images to be wider than the text
720px width works well for text.
The body text has a width between two and three alphabets.
But 720px is not enough for all images.

Change all constraints to 720px to 1000px to adjust the size of the outer container.
Limit title, paragraphs without images, and headlines back to 720px.

This may not be the most elegant long-term solution – more of a quick workaround.

References:
- https://developer.mozilla.org/en-US/docs/Web/CSS/:has
- https://developer.mozilla.org/en-US/docs/Web/CSS/:not